Move common git config options out
This is so that work and personal computers can have their own
user.email and user.name options in the global .gitconfig file. Instead
of committing the entire .gitconfig file, we just commit the common git
options (e.g. aliases and editors) into a .gitconfig.common, then
include the .gitconfig.common file in the .gitconfig file:

    [include]
        path = .gitconfig.common